requirements.txt
discord.py>=2.1.0
aiohttp>=3.7.0
requests>=2.25.1
asyncio>=3.4.3
utils>= 3.4.5
riot-auth>=1.0.3
